Define the StudentExistsByNum method
Now the method does not return true, but accesses the database and returns the correct value.
Added a method for obtaining a student by the number of the record book.

The student is no longer a structure, but a class.
Fields are no longer pointers to char, but arrays.
This is necessary for proper work with the student.