Finish Task54
Add sorting to the list of students from the same group of the same gender for a particular session.
Add a method of printing students from the same group of the same gender with their average mark for all sessions.
Add sorting to the list of students from the same group of the same gender for all sessions.